+++ src/nftSDKFunctions/convert-csv-to-metadata-objects.ts
@@ -17,12 +17,19 @@
  * limitations under the License.
  *
  */
-import { CSVFileReader } from '../csv-file-reader';
+import { readCSVFile } from '../services/csv-file-reader';
 import { MetadataObject } from '../types/csv';
+import { AMOUNT_OF_HEADERS, OMITTED_HEADER_COUNT } from '../utils/constants/csv-constants';
+import { dictionary } from '../utils/constants/dictionary';
 import { prepareMetadataObjectsFromCSVRows } from './prepare-metadata-objects-from-csv-rows';
 
 export const convertCSVToMetadataObjects = async (csvFilePath: string, limit?: number): Promise<MetadataObject[]> => {
-  const csvParsedRows = await CSVFileReader.readCSVFile(csvFilePath, limit);
+  const csvParsedRows = await readCSVFile(csvFilePath, limit);
+
+  if (csvParsedRows.length <= AMOUNT_OF_HEADERS - OMITTED_HEADER_COUNT) {
+    throw new Error(dictionary.validation.csvFileIsEmpty(csvFilePath));
+  }
+
   const metadataObjects = prepareMetadataObjectsFromCSVRows({ csvParsedRows });
 
   return metadataObjects;
