+++ src/nftSDKFunctions/convert-metadata-objects-to-json-files.ts
@@ -20,7 +20,7 @@
 import { JsonMetadataFromCSVInterface } from '../types/json-metadata-from-csv';
 import { Hip412Validator } from '../hip412-validator';
 import { MetadataObject } from '../types/csv';
-import { saveCSVRowsAsJsonFiles } from '../helpers/save-csv-rows-as-json-files';
+import { saveMetadataObjectsAsJsonFiles } from '../helpers/save-metadata-object-as-json-files';
 
 export const convertMetadataObjectsToJsonFiles = async ({
   metadataObjects,
@@ -35,7 +35,7 @@ export const convertMetadataObjectsToJsonFiles = async ({
 
   if (isValid) {
     const objectsToProcess = limit !== undefined && limit < metadataObjects.length ? metadataObjects.slice(0, limit) : metadataObjects;
-    saveCSVRowsAsJsonFiles(objectsToProcess, savedJsonFilesLocation);
+    saveMetadataObjectsAsJsonFiles(objectsToProcess, savedJsonFilesLocation);
   }
 
   return {
