+++ src/nftSDKFunctions/prepare-metadata-objects-from-csv-rows.ts
@@ -1,11 +1,29 @@
-import { JsonMetadataFromCSVConverter } from '../services/json-metadata-from-csv-converter';
+/*-
+ *
+ * Hedera NFT Utilities
+ *
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+import { parseCSVRowsToMetadataObjects } from '../services/json-metadata-from-csv-converter';
 import { CSVRow } from '../types/csv';
 import { ATTRIBUTES, PROPERTIES } from '../utils/constants/csv-constants';
 
-export const prepareMetadataObjectsFromCSVRows = ({ csvParsedRows, csvFilePath }: { csvParsedRows: CSVRow[]; csvFilePath?: string }) => {
-  const metadataObjects = JsonMetadataFromCSVConverter.parseCSVRowsToMetadataObjects({
+export const prepareMetadataObjectsFromCSVRows = ({ csvParsedRows }: { csvParsedRows: CSVRow[] }) => {
+  const metadataObjects = parseCSVRowsToMetadataObjects({
     csvParsedRows,
-    csvFilePath,
     headerAttributes: ATTRIBUTES,
     headerProperties: PROPERTIES,
   });
