+++ src/test/__mocks__/csv/csv-example-with-headers-only.csv
@@ -1,0 +1,2 @@
+name,creator,creatorDID,description,image,type,properties,,attributes,,
+--,--,--,--,--,--,external_url,url,color,hasPipe,stamina
